% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iucn_data.R
\name{get_iucn_data}
\alias{get_iucn_data}
\title{Retrieve full information from iucn_data based on species names}
\usage{
get_iucn_data(species_names)
}
\arguments{
\item{species_names}{A character vector of species names to search for in the \code{iucn_data}.}
}
\value{
A data.table with full information for each species found in \code{iucn_data}. If a species is not found, it will return NA for the respective fields.
}
\description{
This function takes a vector of species names, searches for them in the \code{iucn_data}
dataset, and returns full information about the matched species.
}
\examples{
x <- c("Cedrela odorata", "Persea americana",
       "Panthera uncia", "Lynx lynx", "Ara militaris", "Zonotrichia capencis")
get_iucn_data(x)

}
